+++ packages/adapter-interfaces/src/types/Application.ts
@@ -25,6 +25,7 @@ export const applicationInsertSchema = z.object({
   name: z.string(),
   callbacks: z
     .array(z.string())
+    .optional()
     .openapi({
       description:
         "Comma-separated list of URLs whitelisted to use as a callback to the client after authentication.",
@@ -32,6 +33,7 @@ export const applicationInsertSchema = z.object({
     .default([]),
   allowed_origins: z
     .array(z.string())
+    .optional()
     .openapi({
       description:
         "Comma-separated list of URLs allowed to make requests from JavaScript to Auth0 API (typically used with CORS). By default, all your callback URLs will be allowed. This field allows you to enter other origins if necessary. You can also use wildcards at the subdomain level. Query strings and hash information are not taken into account when validating these URLs.",
@@ -39,6 +41,7 @@ export const applicationInsertSchema = z.object({
     .default([]),
   web_origins: z
     .array(z.string())
+    .optional()
     .openapi({
       description:
         "Comma-separated list of allowed origins for use with Cross-Origin Authentication, Device Flow, and web message response mode.",
@@ -46,6 +49,7 @@ export const applicationInsertSchema = z.object({
     .default([]),
   allowed_logout_urls: z
     .array(z.string())
+    .optional()
     .openapi({
       description:
         "Comma-separated list of URLs that are valid to redirect to after logout from Auth0. Wildcards are allowed for subdomains.",
@@ -53,6 +57,7 @@ export const applicationInsertSchema = z.object({
     .default([]),
   allowed_clients: z
     .array(z.string())
+    .optional()
     .openapi({
       description:
         "Ids of clients that will be allowed to perform delegation requests. Clients that will be allowed to make delegation request. By default, all your clients will be allowed. This field allows you to specify specific clients",
@@ -62,6 +67,7 @@ export const applicationInsertSchema = z.object({
     .object({
       samlp: samlpAddon.optional(),
     })
+    .optional()
     .openapi({
       description:
         "Addons associated with the client. The key is the addon's package name and the value is an object with the configuration for the addon.",
@@ -69,13 +75,17 @@ export const applicationInsertSchema = z.object({
     .default({}),
   email_validation: z
     .enum(["enabled", "disabled", "enforced"])
+    .optional()
     .default("enforced")
     .openapi({
       description:
         "Defines if it possible to sign in with an unverified email and if verification emails will be sent. This is not available in auth0",
     }),
-  client_secret: z.string().default(() => nanoid()),
-  disable_sign_ups: z.boolean().default(false).openapi({
+  client_secret: z
+    .string()
+    .optional()
+    .default(() => nanoid()),
+  disable_sign_ups: z.boolean().optional().default(false).openapi({
     description:
       "Prevents users from signing up using the hosted login page. This is not available in auth0",
   }),