+++ packages/adapter-interfaces/src/types/Connection.ts
@@ -30,9 +30,10 @@ export const connectionInsertSchema = z.object({
       app_secret: z.string().optional(),
       scope: z.string().optional(),
     })
-    .optional(),
-  enabled_clients: z.array(z.string()).optional(),
-  authorization_endpoint: z.string().optional(),
+    .optional()
+    .default({}),
+  enabled_clients: z.array(z.string()).optional().default([]),
+  authorization_endpoint: z.string().optional().default(""),
   response_type: z.custom<AuthorizationResponseType>().optional(),
   response_mode: z.custom<AuthorizationResponseMode>().optional(),