+++ packages/authhero/CHANGELOG.md
@@ -1,5 +1,12 @@
 # authhero
 
+## 0.2.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.15.3
+
 ## 0.2.25
 
 ### Patch Changes